requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyter
google-cloud-bigquery
pyarrow